MXS-4631 Fix a bug in find_last_gtid_list()
This fixes the case where a completed file is read,
meaning that there is a STOP or ROTATE at the end of
the file. A ROTATE's next_event_pos points back to itself.

This will allow a reader to read files using
find_last_gtid_list(), waiting for a certain gtid to
have been recorded.